updated get config api call to update everything
