+++ src/frontend/src/controllers/API/queries/config/use-get-config.ts
@@ -1,3 +1,5 @@
+import useFlowsManagerStore from "@/stores/flowsManagerStore";
+import axios from "axios";
 import { useQueryFunctionType } from "../../../../types/api";
 import { api } from "../../api";
 import { getURL } from "../../helpers/constants";
@@ -10,18 +12,35 @@ export interface ConfigResponse {
   health_check_max_retries: number;
 }
 
-export const useGetConfigQuery: useQueryFunctionType<
-  undefined,
-  ConfigResponse
-> = (options) => {
+export const useGetConfig: useQueryFunctionType<undefined, undefined> = (
+  options,
+) => {
+  const setAutoSaving = useFlowsManagerStore((state) => state.setAutoSaving);
+  const setAutoSavingInterval = useFlowsManagerStore(
+    (state) => state.setAutoSavingInterval,
+  );
+  const setHealthCheckMaxRetries = useFlowsManagerStore(
+    (state) => state.setHealthCheckMaxRetries,
+  );
+
   const { query } = UseRequestProcessor();
 
   const getConfigFn = async () => {
     const response = await api.get<ConfigResponse>(`${getURL("CONFIG")}`);
-    return response["data"];
+    const data = response["data"];
+    if (data) {
+      const timeoutInMilliseconds = data.frontend_timeout
+        ? data.frontend_timeout * 1000
+        : 30000;
+      axios.defaults.baseURL = "";
+      axios.defaults.timeout = timeoutInMilliseconds;
+      setAutoSaving(data.auto_saving);
+      setAutoSavingInterval(data.auto_saving_interval);
+      setHealthCheckMaxRetries(data.health_check_max_retries);
+    }
   };
 
-  const queryResult = query(["useGetConfigQuery"], getConfigFn, options);
+  const queryResult = query(["useGetConfig"], getConfigFn, options);
 
   return queryResult;
 };
